auth: Move constructors back to their structs
There isn't a lot of conventions that we reliably stick to - keeping
constructors in the same file as their structs is one of them, so I'm
going to put these back.
