+++ pkg/server/auth/init.go
@@ -3,15 +3,10 @@ package auth
 import (
 	"context"
 	"fmt"
-	"net/http"
-	"net/url"
-	"time"
 
-	"github.com/coreos/go-oidc/v3/oidc"
 	"github.com/go-logr/logr"
 	"github.com/weaveworks/weave-gitops/api/v1alpha1"
 	"github.com/weaveworks/weave-gitops/core/logger"
-	"github.com/weaveworks/weave-gitops/pkg/featureflags"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
@@ -79,82 +74,4 @@ func InitAuthServer(ctx context.Context, log logr.Logger, rawKubernetesClient ct
 	}
 
 	return authServer, err
-}
-
-func NewOIDCConfigFromSecret(secret corev1.Secret) OIDCConfig {
-	cfg := OIDCConfig{
-		IssuerURL:    string(secret.Data["issuerURL"]),
-		ClientID:     string(secret.Data["clientID"]),
-		ClientSecret: string(secret.Data["clientSecret"]),
-		RedirectURL:  string(secret.Data["redirectURL"]),
-	}
-
-	tokenDuration, err := time.ParseDuration(string(secret.Data["tokenDuration"]))
-	if err != nil {
-		tokenDuration = defaultCookieDuration
-	}
-
-	cfg.TokenDuration = tokenDuration
-
-	return cfg
-}
-
-func NewAuthServerConfig(log logr.Logger, oidcCfg OIDCConfig, kubernetesClient ctrlclient.Client, tsv TokenSignerVerifier, authMethods map[AuthMethod]bool) (AuthConfig, error) {
-	if authMethods[OIDC] {
-		if _, err := url.Parse(oidcCfg.IssuerURL); err != nil {
-			return AuthConfig{}, fmt.Errorf("invalid issuer URL: %w", err)
-		}
-
-		if _, err := url.Parse(oidcCfg.RedirectURL); err != nil {
-			return AuthConfig{}, fmt.Errorf("invalid redirect URL: %w", err)
-		}
-	}
-
-	return AuthConfig{
-		Log:                 log.WithName("auth-server"),
-		client:              http.DefaultClient,
-		kubernetesClient:    kubernetesClient,
-		tokenSignerVerifier: tsv,
-		config:              oidcCfg,
-		authMethods:         authMethods,
-	}, nil
-}
-
-// NewAuthServer creates a new AuthServer object.
-func NewAuthServer(ctx context.Context, cfg AuthConfig) (*AuthServer, error) {
-	if cfg.authMethods[UserAccount] {
-		var secret corev1.Secret
-		err := cfg.kubernetesClient.Get(ctx, client.ObjectKey{
-			Namespace: v1alpha1.DefaultNamespace,
-			Name:      ClusterUserAuthSecretName,
-		}, &secret)
-
-		if err != nil {
-			return nil, fmt.Errorf("Could not get secret for cluster user, %w", err)
-		} else {
-			featureflags.Set(FeatureFlagClusterUser, FeatureFlagSet)
-		}
-	} else {
-		featureflags.Set(FeatureFlagClusterUser, "false")
-	}
-
-	var provider *oidc.Provider
-
-	if cfg.config.IssuerURL == "" {
-		featureflags.Set(FeatureFlagOIDCAuth, "false")
-	} else if cfg.authMethods[OIDC] {
-		var err error
-
-		provider, err = oidc.NewProvider(ctx, cfg.config.IssuerURL)
-		if err != nil {
-			return nil, fmt.Errorf("could not create provider: %w", err)
-		}
-		featureflags.Set(FeatureFlagOIDCAuth, FeatureFlagSet)
-	}
-
-	if featureflags.Get(FeatureFlagOIDCAuth) != FeatureFlagSet && featureflags.Get(FeatureFlagClusterUser) != FeatureFlagSet {
-		return nil, fmt.Errorf("Neither OIDC auth or local auth enabled, can't start")
-	}
-
-	return &AuthServer{cfg, provider}, nil
 }
